components/udToggleView/udToggleView.definition.spec.ts
import { UdToggleViewComponentDefinition, UdToggleViewComponentState } from './udToggleView.definition'
import { generateSchemaAndValidate } from '../../../utilities'

describe('UD Toggle View Contract Tests', function () {
  it('should ensure the contract is clear - component definition', function () {
    const { schema } = generateSchemaAndValidate(UdToggleViewComponentDefinition)

    expect(schema.required).toMatchSnapshot()
    expect(schema.properties).toMatchSnapshot()
    expect(schema.definitions).toMatchSnapshot()
  })

  it('should ensure the contract is clear - component state', function () {
    const { schema } = generateSchemaAndValidate(UdToggleViewComponentState)

    expect(schema.required).toMatchSnapshot()
    expect(schema.properties).toMatchSnapshot()
    expect(schema.definitions).toMatchSnapshot()
  })
})
